move subhalo potential
